feat: :goal_net: Tratativa caso relacionamento exista ou não nos casos de uso de atribuição e desatribuição de tarefa

+++ src/modules/tasks/use-cases/unassign-task-from-user.ts
@@ -4,6 +4,7 @@ import { UsersRepository } from "@modules/users/repositories/users.repository";
 
 import { TasksUsersRepository } from "../repositories/tasks-users.repository";
 import { TasksRepository } from "../repositories/tasks.repository";
+import { NotAssigned } from "./errors/not-assigned";
 import { TaskNotFound } from "./errors/task-not-found";
 import { UserNotFound } from "./errors/user-not-found";
 
@@ -31,6 +32,15 @@ export class UnassignTaskFromUserUseCase {
 
     if (!task) {
       throw new TaskNotFound();
+    }
+
+    const assigned = await this.tasksUsersRepository.findByTaskIdAndUserId(
+      userId,
+      taskId
+    );
+
+    if (!assigned) {
+      throw new NotAssigned();
     }
 
     await this.tasksUsersRepository.delete(userId, taskId);
